Allow specifying options to aioredis create_pool
